This is a script to check and update user only chrome installs
I built this script because users where installing chrome with out Admin permissions, there for it would install in the %userappdata%\Google\Chrome\Application.
But the issue is when they stop using it the application will not update.

This script is there for comprmise, because our admins won't let me delete when they have been active with in the last year. So I keep the chrome data for the user good, and just update the application.