Update user-cleanup-windows.ps1
Fixed some issues with the loops and the old check var.

I added a better check, if the IconCache.db is not there look at the \local\microsoft\widnows Direcotry. This seems to line up with logs from our AD for log in times. So I think it will be a good back up.