Monitor-ID.ps1
This is a script I got from https://amgeneral.wordpress.com/2021/07/13/powershell-list-monitors-and-connection-types/

I've been using it to ID monitors on windows 10. I've been needing to do this with Dell Equipment because the Genric PNP Montior Drive in Windows 10 and 11 has been causing issue wtih Blinking randomly. So I use this to find the Display Model, and then install the Drivers for that monitor.